export side menu js function to own file
