+++ app/javascript/packs/application.js
@@ -11,21 +11,11 @@ require("channels")
 import { getLocation } from '../plugins/retrieve_user_location';
 import { itemNotification } from '../plugins/notification';
 import { init_flatpickr } from '../plugins/init_flatpickr';
+import { toggleSideMenu } from '../plugins/side_menu';
 
 document.addEventListener('turbolinks:load', () => {
   getLocation();
   itemNotification();
   init_flatpickr();
-  
-  const menuModal = document.querySelector('.menu-modal');
-  const burgerMenu = document.querySelector('.burger-menu');
-  burgerMenu.addEventListener('click', () => {
-    menuModal.style.transitionDuration = "0.5s";
-    menuModal.style.right = '0';
-  });
-  
-  const closeMenu = document.querySelector('.close-menu');
-  closeMenu.addEventListener('click', () => {
-    menuModal.style.right = '-300px';
-  });
+  toggleSideMenu();
 });
